[FLINK-17681] Correct type when converting Row, Map, Array to RexNodes
This commit fixes the expression type when converting from a
CallExpession to RexNode for Row, Map, Array constructors.

It also adds support for LegacyTypeInformationType in FlinkTypeFactory
as it might end up there from the type inference.

Lastly it makes the Row, Map, Array ctors return not null type, as the
ctors can not produce a null.
